+++ tests/test_convert.py
@@ -5,6 +5,8 @@ import os
 
 import mesoscopy
 import mesoscopy.convert as conv
+import mesoscopy.convert.hdf5 as conv_h5
+import mesoscopy.convert.video as conv_vid
 import mesoscopy.convert.metadata as mtd
 
 from pynwb import NWBHDF5IO
@@ -28,7 +30,7 @@ def test_metadata_json(meta_json):
 
 def test_convert_h5_linkonly(raw_h5, output_dir):
     expected_outpath = output_dir + os.sep + raw_h5.split(os.sep)[-1].replace(".h5", ".nwb")
-    nwb_outpath = conv.convert(raw_h5, out_dir=output_dir, link_only=True)
+    nwb_outpath = conv_h5.to_nwb(raw_h5, out_dir=output_dir, link_only=True)
     assert expected_outpath == nwb_outpath
     assert os.path.isfile(nwb_outpath)
     assert os.path.getsize(raw_h5) > os.path.getsize(nwb_outpath)
@@ -36,14 +38,14 @@ def test_convert_h5_linkonly(raw_h5, output_dir):
 
 def test_convert_h5_eagercopy(raw_h5, output_dir):
     expected_outpath = output_dir + os.sep + raw_h5.split(os.sep)[-1].replace(".h5", ".nwb")
-    nwb_outpath = conv.convert(raw_h5, out_dir=output_dir)
+    nwb_outpath = conv_h5.to_nwb(raw_h5, out_dir=output_dir)
     assert expected_outpath == nwb_outpath
     assert os.path.isfile(nwb_outpath)
     assert os.path.getsize(raw_h5) <= os.path.getsize(nwb_outpath)
 
 
 def test_convert_h5_metadata_yaml(raw_h5, output_dir, meta_yaml):
-    nwb_outpath = conv.convert(raw_h5, out_dir=output_dir, meta_path=meta_yaml)
+    nwb_outpath = conv_h5.to_nwb(raw_h5, out_dir=output_dir, meta_path=meta_yaml)
 
     io = NWBHDF5IO(nwb_outpath, mode="r")
     nwbfile = io.read()
@@ -51,7 +53,7 @@ def test_convert_h5_metadata_yaml(raw_h5, output_dir, meta_yaml):
 
 
 def test_convert_h5_metadata_json(raw_h5, output_dir, meta_json):
-    nwb_outpath = conv.convert(raw_h5, out_dir=output_dir, meta_path=meta_json)
+    nwb_outpath = conv_h5.to_nwb(raw_h5, out_dir=output_dir, meta_path=meta_json)
 
     io = NWBHDF5IO(nwb_outpath, mode="r")
     nwbfile = io.read()
@@ -59,7 +61,7 @@ def test_convert_h5_metadata_json(raw_h5, output_dir, meta_json):
 
 
 def test_convert_h5_metadata_args(raw_h5, output_dir):
-    nwb_outpath = conv.convert(
+    nwb_outpath = conv_h5.to_nwb(
         raw_h5,
         out_dir=output_dir,
         subject_id="sometest",
@@ -84,11 +86,26 @@ def test_convert_h5_metadata_args(raw_h5, output_dir):
 def test_convert_h5_metadata_args_file_mixed(raw_h5, output_dir): ...
 
 
-def test_cmd_parity(raw_h5, output_dir):
-    """Check if click convert cmd caller works"""
+def test_h5_cmd_parity(raw_h5, output_dir):
+    """Check if click convert hdf5 cmd caller works"""
     runner = CliRunner()
     result = runner.invoke(
         mesoscopy.cli,
-        args=f"convert --out-dir {output_dir} {raw_h5}",
+        args=f"convert h5 --out-dir {output_dir} {raw_h5}",
     )
     assert result.exit_code == 0
+
+
+def test_avi_h5_converter(raw_avi, output_dir):
+    h5_outpath = conv_vid.to_hdf5(raw_avi, out_dir=output_dir)
+    assert os.path.isfile(h5_outpath)
+
+
+def test_avi_h5_converter_with_timestamps(raw_avi, raw_timestamps, output_dir):
+    h5_outpath = conv_vid.to_hdf5(raw_avi, output_dir, ts_path=raw_timestamps)
+    assert os.path.isfile(h5_outpath)
+
+
+def test_avi_nwb_converter(raw_avi, output_dir):
+    h5_outpath = conv_vid.to_nwb(raw_avi, out_dir=output_dir)
+    assert os.path.isfile(h5_outpath)