+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Versions follow [Semantic Versioning](https://semver.org) (`<major>.<minor>.<pat
 ### Added
 
 - Registration GUI with Napari for identifying anatomical landmarks.
-- Conversion CLI command to create NWB files from raw HDF5 recordings.
+- Conversion CLI command to create NWB files from raw HDF5 and video recordings (via `convert h5` and `convert video`, respectively).
 - Mesoscopy NWB file inspection with the new `inspect` command.
 
 ### Changed